Create pfTracker.cs
Got frustraited for the amount of errors, I stumbuled upon, so I will leve it as it is for now, treating it as a base code. Need to fix the math, and if possible making it able to save multiple files, and allowing the user to name the files. Maybe later add filtering for csv //I doubt it